moved AliquotImage back to aliquots package for backwards compatibility
